♻️ Implements cleaner APIs
