+++ .env.local.example
@@ -4,4 +4,3 @@ MAILGUN_DOMAIN=""
 NEXT_PUBLIC_AUTH_COOKIE_NAME="links_auth_cookie"
 NEXT_PUBLIC_USER_COOKIE_NAME="links_user_cookie"
 SECRET_COOKIE_PASSWORD_1="00000000000000000000000000000000"
-TAILWIND_DISABLE_TOUCH="true"
