+++ lib/api.js
@@ -1,0 +1,5 @@
+import connect from 'next-connect'
+import { session } from './session'
+
+export const handler = connect()
+handler.use(session)
