+++ package.json
@@ -3,11 +3,12 @@
     "autosize": "5.0.0",
     "@prisma/client": "2.23.0",
     "cheerio": "1.0.0-rc.9",
-    "formik": "2.2.8",
+    "formik": "2.2.9",
     "js-cookie": "2.2.1",
     "lodash": "4.17.21",
     "mailgun-js": "0.22.0",
-    "next": "10.2.3",
+    "next": "10.2.4-canary.4",
+    "next-connect": "0.10.1",
     "next-iron-session": "4.1.13",
     "nprogress": "0.2.0",
     "react": "17.0.2",
@@ -16,12 +17,12 @@
     "yup": "0.32.9"
   },
   "devDependencies": {
-    "@tailwindcss/forms": "0.3.2",
+    "@tailwindcss/forms": "0.3.3",
     "autoprefixer": "10.2.6",
     "postcss": "8.3.0",
     "prisma": "2.23.0",
     "prop-types": "15.7.2",
-    "tailwindcss": "2.2.0-canary.11"
+    "tailwindcss": "2.2.0-canary.13"
   },
   "license": "UNLICENSED",
   "prettier": {
